+++ packages/backend/src/scheduler/SchedulerTask.ts
@@ -21,7 +21,6 @@ import {
     NotificationFrequency,
     NotificationPayloadBase,
     QueryExecutionContext,
-    QueryHistoryStatus,
     ReadFileError,
     RenameResourcesPayload,
     ReplaceCustomFields,
@@ -42,6 +41,7 @@ import {
     SessionUser,
     SlackInstallationNotFoundError,
     SlackNotificationPayload,
+    SqlChart,
     SqlRunnerPayload,
     SqlRunnerPivotQueryPayload,
     ThresholdOperator,
@@ -56,9 +56,12 @@ import {
     convertReplaceableFieldMatchMapToReplaceCustomFields,
     formatRows,
     friendlyName,
+    getColumnOrderFromVizTableConfig,
     getCustomLabelsFromTableConfig,
+    getCustomLabelsFromVizTableConfig,
     getErrorMessage,
     getFulfilledValues,
+    getHiddenFieldsFromVizTableConfig,
     getHiddenTableFields,
     getHumanReadableCronExpression,
     getItemMap,
@@ -77,10 +80,10 @@ import {
     isSchedulerGsheetsOptions,
     isSchedulerImageOptions,
     isTableChartConfig,
+    isVizTableConfig,
     operatorActionValue,
     pivotResultsAsCsv,
     setUuidParam,
-    sleep,
 } from '@lightdash/common';
 import fs from 'fs/promises';
 import { nanoid } from 'nanoid';
@@ -509,6 +512,8 @@ export default class SchedulerTask {
                         const csvForChartPromises =
                             chartTileUuidsWithChartUuids.map(
                                 async ({ chartUuid }) => {
+                                    const chartLimit =
+                                        getSchedulerCsvLimit(csvOptions);
                                     const query =
                                         await this.asyncQueryService.executeAsyncDashboardChartQuery(
                                             {
@@ -521,7 +526,7 @@ export default class SchedulerTask {
                                                 dashboardUuid,
                                                 dashboardFilters,
                                                 dashboardSorts: [],
-                                                // todo: support limit arg
+                                                limit: chartLimit,
                                             },
                                         );
                                     const chart =
@@ -601,11 +606,30 @@ export default class SchedulerTask {
                                             type: DownloadFileType.XLSX,
                                             onlyRaw:
                                                 csvOptions?.formatted === false,
-                                            // todo: support this in next pr
-                                            // customLabels: getCustomLabelsFromTableConfig(chart.chartConfig.config),
-                                            // hiddenFields: getHiddenTableFields(chart.chartConfig),
-                                            // pivotConfig: getPivotConfig(chart),
-                                            // columnOrder: chart.tableConfig.columnOrder,
+                                            customLabels:
+                                                getCustomLabelsFromVizTableConfig(
+                                                    isVizTableConfig(
+                                                        chart.config,
+                                                    )
+                                                        ? chart.config
+                                                        : undefined,
+                                                ),
+                                            hiddenFields:
+                                                getHiddenFieldsFromVizTableConfig(
+                                                    isVizTableConfig(
+                                                        chart.config,
+                                                    )
+                                                        ? chart.config
+                                                        : undefined,
+                                                ),
+                                            columnOrder:
+                                                getColumnOrderFromVizTableConfig(
+                                                    isVizTableConfig(
+                                                        chart.config,
+                                                    )
+                                                        ? chart.config
+                                                        : undefined,
+                                                ),
                                         },
                                     );
                                 return {